+++ src/org/opendatakit/briefcase/reused/http/Http.java
@@ -16,6 +16,8 @@
 
 package org.opendatakit.briefcase.reused.http;
 
+import org.opendatakit.briefcase.reused.http.response.Response;
+
 /**
  * This interface has Briefcase's HTTP API to interact with external services
  */
